+++ clients/derivativestradingcoinfutures/src/restapi/models/model_modify_multiple_orders_response_inner.go
@@ -27,6 +27,8 @@ type ModifyMultipleOrdersResponseInner struct {
 	Status *string `json:"status,omitempty"`
 	// Client order ID.
 	ClientOrderId *string `json:"clientOrderId,omitempty"`
+	// user-defined modification identifier, only returned if provided in the request
+	ModifyId *int64 `json:"modifyId,omitempty"`
 	// Latest token price.
 	Price *string `json:"price,omitempty"`
 	// Original order quantity
@@ -245,6 +247,38 @@ func (o *ModifyMultipleOrdersResponseInner) HasClientOrderId() bool {
 // SetClientOrderId gets a reference to the given string and assigns it to the ClientOrderId field.
 func (o *ModifyMultipleOrdersResponseInner) SetClientOrderId(v string) {
 	o.ClientOrderId = &v
+}
+
+// GetModifyId returns the ModifyId field value if set, zero value otherwise.
+func (o *ModifyMultipleOrdersResponseInner) GetModifyId() int64 {
+	if o == nil || common.IsNil(o.ModifyId) {
+		var ret int64
+		return ret
+	}
+	return *o.ModifyId
+}
+
+// GetModifyIdOk returns a tuple with the ModifyId field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *ModifyMultipleOrdersResponseInner) GetModifyIdOk() (*int64, bool) {
+	if o == nil || common.IsNil(o.ModifyId) {
+		return nil, false
+	}
+	return o.ModifyId, true
+}
+
+// HasModifyId returns a boolean if a field has been set.
+func (o *ModifyMultipleOrdersResponseInner) HasModifyId() bool {
+	if o != nil && !common.IsNil(o.ModifyId) {
+		return true
+	}
+
+	return false
+}
+
+// SetModifyId gets a reference to the given int64 and assigns it to the ModifyId field.
+func (o *ModifyMultipleOrdersResponseInner) SetModifyId(v int64) {
+	o.ModifyId = &v
 }
 
 // GetPrice returns the Price field value if set, zero value otherwise.
@@ -880,6 +914,9 @@ func (o ModifyMultipleOrdersResponseInner) ToMap() (map[string]interface{}, erro
 	if !common.IsNil(o.ClientOrderId) {
 		toSerialize["clientOrderId"] = o.ClientOrderId
 	}
+	if !common.IsNil(o.ModifyId) {
+		toSerialize["modifyId"] = o.ModifyId
+	}
 	if !common.IsNil(o.Price) {
 		toSerialize["price"] = o.Price
 	}
@@ -964,6 +1001,7 @@ func (o *ModifyMultipleOrdersResponseInner) UnmarshalJSON(data []byte) (err erro
 		delete(additionalProperties, "pair")
 		delete(additionalProperties, "status")
 		delete(additionalProperties, "clientOrderId")
+		delete(additionalProperties, "modifyId")
 		delete(additionalProperties, "price")
 		delete(additionalProperties, "origQty")
 		delete(additionalProperties, "executedQty")