+++ clients/derivativestradingportfoliomargin/src/restapi/models/model_modify_um_order_response.go
@@ -25,6 +25,8 @@ type ModifyUmOrderResponse struct {
 	Status *string `json:"status,omitempty"`
 	// Client Order ID.
 	ClientOrderId *string `json:"clientOrderId,omitempty"`
+	// user-defined modification identifier, only returned if provided in the request
+	ModifyId *int64 `json:"modifyId,omitempty"`
 	// Price.
 	Price *string `json:"price,omitempty"`
 	// Avg Price.
@@ -205,6 +207,38 @@ func (o *ModifyUmOrderResponse) HasClientOrderId() bool {
 // SetClientOrderId gets a reference to the given string and assigns it to the ClientOrderId field.
 func (o *ModifyUmOrderResponse) SetClientOrderId(v string) {
 	o.ClientOrderId = &v
+}
+
+// GetModifyId returns the ModifyId field value if set, zero value otherwise.
+func (o *ModifyUmOrderResponse) GetModifyId() int64 {
+	if o == nil || common.IsNil(o.ModifyId) {
+		var ret int64
+		return ret
+	}
+	return *o.ModifyId
+}
+
+// GetModifyIdOk returns a tuple with the ModifyId field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *ModifyUmOrderResponse) GetModifyIdOk() (*int64, bool) {
+	if o == nil || common.IsNil(o.ModifyId) {
+		return nil, false
+	}
+	return o.ModifyId, true
+}
+
+// HasModifyId returns a boolean if a field has been set.
+func (o *ModifyUmOrderResponse) HasModifyId() bool {
+	if o != nil && !common.IsNil(o.ModifyId) {
+		return true
+	}
+
+	return false
+}
+
+// SetModifyId gets a reference to the given int64 and assigns it to the ModifyId field.
+func (o *ModifyUmOrderResponse) SetModifyId(v int64) {
+	o.ModifyId = &v
 }
 
 // GetPrice returns the Price field value if set, zero value otherwise.
@@ -741,6 +775,9 @@ func (o ModifyUmOrderResponse) ToMap() (map[string]interface{}, error) {
 	if !common.IsNil(o.ClientOrderId) {
 		toSerialize["clientOrderId"] = o.ClientOrderId
 	}
+	if !common.IsNil(o.ModifyId) {
+		toSerialize["modifyId"] = o.ModifyId
+	}
 	if !common.IsNil(o.Price) {
 		toSerialize["price"] = o.Price
 	}
@@ -815,6 +852,7 @@ func (o *ModifyUmOrderResponse) UnmarshalJSON(data []byte) (err error) {
 		delete(additionalProperties, "symbol")
 		delete(additionalProperties, "status")
 		delete(additionalProperties, "clientOrderId")
+		delete(additionalProperties, "modifyId")
 		delete(additionalProperties, "price")
 		delete(additionalProperties, "avgPrice")
 		delete(additionalProperties, "origQty")